Create and move code to operations module
